Functions/server/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
exceptiongroup==1.2.0
fastapi==0.104.1
firebase-admin==6.3.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
httplib2==0.22.0
idna==3.6
msgpack==1.0.7
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyparsing==3.1.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
